requirements.txt
Flask==3.0.3
flask-restx==1.3.0
Flask-Mail==0.9.1
python-dotenv==1.0.1
pymongo==4.8.0
Werkzeug==3.0.3
itsdangerous==2.2.0
